Almost fully functioning Login system based on MVVM (need isEnable)
