+++ SuppLocals/SuppLocals/Views/AccountViews/ForgotPassword.xaml.cs
@@ -41,7 +41,7 @@ namespace SuppLocals.Views
 
         private void BackButtonClick(object sender, RoutedEventArgs e)
         {
-            var login = new Login();
+            var login = new LoginSignupWindow();
             login.Show();
             Close();
         }
