--- SuppLocals/SuppLocals/Views/AccountViews/Login.xaml
@@ -1,163 +1,0 @@
-﻿<Window x:Class="SuppLocals.Views.Login"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SuppLocals"
-        mc:Ignorable="d"
-        Title="Login" Height="500" Width="400"
-        AllowsTransparency="true" Background="Transparent"
-        WindowStyle="none" ResizeMode="NoResize"
-        WindowStartupLocation="CenterScreen"
-        MouseDown="Window_MouseDown">
-
-
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-
-        <ContentControl Grid.Row="2" Grid.Column="0" Grid.RowSpan="2" Content="{Binding}"/>
-
-        <Border CornerRadius="10" Grid.RowSpan="2">
-            <Border.Background>
-                <LinearGradientBrush>
-                    <GradientStop Color="#CCBA8B" Offset="0.0"/>
-                    <GradientStop Color="#CCBA8B" Offset="1"/>
-                </LinearGradientBrush>
-            </Border.Background>
-        </Border>
-
-        <StackPanel VerticalAlignment="Center">
-            <Image Source="/Assets/title2.png" Width="300" Visibility="Visible"/>
-        </StackPanel>
-
-
-        <StackPanel Grid.Row="1">
-            <StackPanel Orientation="Horizontal">
-                <Label  Foreground="Black" 
-                        Content="Username:"
-                        FontFamily="Tahoma"
-                        FontSize="20"
-                        FontWeight="Bold"
-                        Margin="70 -10 0 0 "/>
-
-                <TextBox    Name="Username"
-                        FontFamily="Tahoma"
-                        FontWeight="Bold"
-                        FontSize="20"
-                        Foreground="Black"
-                        Background="Transparent"
-                        BorderThickness="0"
-                        Width="235"
-                        HorizontalContentAlignment="Left"
-                        Opacity="1"
-
-                        TextChanged="UsernameTextChangedEventHandler"
-                        Margin="-110 25 0 0 "
-                        MaxLength="12"
-                        Text="{Binding Username, 
-                        UpdateSourceTrigger=PropertyChanged, 
-                        ValidatesOnDataErrors=True}"
-                        ToolTip="{Binding ErrorCollection[Username]}">
-                </TextBox>
-
-            </StackPanel>
-            <Border Width="250"
-                    Height="2"
-                    Background="Black"
-                    Opacity="0.5"/>
-
-            <StackPanel Orientation="Horizontal">
-                <Label  Foreground="Black" 
-                        Content="Password:"
-                        FontFamily="Tahoma"
-                        FontSize="20"
-                        FontWeight="Bold"
-                        Margin="70 0 0 0 "/>
-
-                <PasswordBox  local:PasswordHelper.Attach="True"
-                        local:PasswordHelper.Password="{Binding Password,
-                        Mode=TwoWay,  UpdateSourceTrigger=PropertyChanged, 
-                        ValidatesOnDataErrors=True}"
-                        ToolTip="{Binding ErrorCollection[Password]}"
-                        Name="PasswordBox"
-                        MaxLength="12"
-                        FontFamily="Tahoma"
-                        FontWeight="Bold"
-                        Password=""
-                        FontSize="20"
-                        Foreground="Black"
-                        Background="Transparent"
-                        BorderThickness="0"
-                        Width="235"
-                        HorizontalContentAlignment="Left"
-                        Opacity="1"
-                        Height="25"
-                        Margin="-110 35 0 0 "
-                        PasswordChanged="PasswordTextChangedEventHandler"/>
-            </StackPanel>
-            <Border Width="250"
-                    Height="2"
-                    Background="Black"
-                    Opacity="0.5"/>
-
-            <StackPanel Orientation="Horizontal" Margin="100 30 0 0">
-                <Button x:Name="loginBtn" Width="100" Height="30" 
-                        Content="Login" Click="LogInBtnClick" IsEnabled="False"/>
-                <Button Width="100" Height="30" Content="Register" Click="SignUpBtnClick"/>
-            </StackPanel>
-
-            <Button Width="100" Height="30" Content="Go to Map" Click="GoToMapClick"/>
-
-            <Label Height="25" Width="110" 
-                   Margin="160,-90,0,0"
-                   Name="forgotPassword" VerticalAlignment="Top">
-                <Hyperlink Click="Hyperlink_Click">
-                    Forgot Password?</Hyperlink>
-            </Label>
-
-        </StackPanel>
-
-        
-        
-    </Grid>
-
-    <Window.Resources>
-        <Style TargetType="{x:Type TextBox}">
-            <Setter Property="Validation.ErrorTemplate">
-                <Setter.Value>
-                    <ControlTemplate>
-                        <DockPanel>
-                            <Grid DockPanel.Dock="Left" Width="16" Height="16" VerticalAlignment="Center" Margin="230 0 0 0">
-                                <Ellipse Width="16" Height="16" Fill="Red"/>
-                                <Ellipse Width="3" Height="8" VerticalAlignment="Top" HorizontalAlignment="Center" Margin="0 2 0 0" Fill="White"/>
-                                <Ellipse Width="2" Height="2" VerticalAlignment="Bottom" HorizontalAlignment="Center" Margin="0 0 0 2" Fill="White"/>
-                            </Grid>
-                        </DockPanel>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-        </Style>
-
-        <Style TargetType="{x:Type PasswordBox}">
-            <Setter Property="Validation.ErrorTemplate">
-                <Setter.Value>
-                    <ControlTemplate>
-                        <DockPanel>
-                            <Grid DockPanel.Dock="Left" Width="16" Height="16" VerticalAlignment="Center" Margin="230 0 0 0">
-                                <Ellipse Width="16" Height="16" Fill="Red"/>
-                                <Ellipse Width="3" Height="8" VerticalAlignment="Top" HorizontalAlignment="Center" Margin="0 2 0 0" Fill="White"/>
-                                <Ellipse Width="2" Height="2" VerticalAlignment="Bottom" HorizontalAlignment="Center" Margin="0 0 0 2" Fill="White"/>
-                            </Grid>
-                        </DockPanel>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-        </Style>
-
-    </Window.Resources>
-    
-
-</Window>
