+++ SuppLocals/SuppLocals/Views/AccountViews/LoginView.xaml
@@ -104,21 +104,13 @@
                     Background="Black"
                     Opacity="0.5"/>
 
-        <StackPanel Orientation="Horizontal" Margin="100 10 0 0">
+            <StackPanel Orientation="Horizontal" Margin="100 10 0 0">
                 <Button x:Name="loginBtn" Width="100" Height="30" 
                         Content="Login" 
                          Command="{Binding LoginClick}"/>
+            </StackPanel>
+
         </StackPanel>
-
-
-        <Label Height="25" Width="110" 
-                   Margin="160,10,0,0"
-                   Name="forgotPassword" VerticalAlignment="Top">
-            <Hyperlink >
-                Forgot Password?</Hyperlink>
-        </Label>
-
-    </StackPanel>
 
 
     </Grid>
