+++ SuppLocals/SuppLocals/Views/LoginSignupWindow.xaml.cs
@@ -1,8 +1,10 @@
-﻿using SuppLocals.ViewModels;
+﻿using SuppLocals.Utilities.Helpers;
+using SuppLocals.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Windows;
+using System.Windows.Input;
 
 namespace SuppLocals.Views
 {
@@ -12,17 +14,20 @@ namespace SuppLocals.Views
         {
             InitializeComponent();
             DataContext = new LoginVM();
+            CloseWindow.WinObject = (Window)this;
 
         }
 
         private void LoginClick(object sender, RoutedEventArgs e)
         {
             DataContext = new LoginVM();
+            Labeliukas.Visibility = Visibility.Visible;
         }
 
         private void SignupClick(object sender, RoutedEventArgs e)
         {
             DataContext = new SignupVM();
+            Labeliukas.Visibility = Visibility.Hidden;
             
         }
 
@@ -31,6 +36,21 @@ namespace SuppLocals.Views
             var start = new StartWindow();
             start.Show();
             this.Close();
+        }
+
+        private void Hyperlink_Click(object sender, RoutedEventArgs e)
+        {
+            var forgotPassword = new ForgotPassword();
+            forgotPassword.Show();
+            this.Close();
+        }
+
+        private void Window_MouseDown(object sender, MouseButtonEventArgs e)
+        {
+            if (e.LeftButton == MouseButtonState.Pressed)
+            {
+                DragMove();
+            }
         }
     }
 }
